requirements.txt
asgiref==3.9.1
Django==5.2.6
pyserial==3.5
sqlparse==0.5.3
tzdata==2025.2
